requirements.txt
streamlit
pandas
plotly
openpyxl
python-docx
Pillow
kaleido==0.2.1
matplotlib
PyGithub